pyproject.toml
[tool.poetry]
name = "litestar-habittracker"
version = "0.1.0"
description = "API for habit tracker app"
authors = ["darialissi"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.12"
litestar = {extras = ["standard"], version = "^2.12.1"}
litestar-asyncpg = "^0.2.0"
pydantic = "^2.9.2"
bcrypt = "^4.2.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
pydantic-settings = "^2.6.1"
advanced-alchemy = "^0.23.1"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
Flake8-pyproject = "^1.2.3"
pytest-asyncio = "^0.24.0"
pytest-dotenv = "^0.5.2"


[tool.poetry.group.mcp.dependencies]
fastmcp = ">=3.0.0b2"

[tool.pytest.ini_options]
env_override_existing_values = 1
env_files = [".env-test"]
pythonpath = [".", "src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = "ignore::DeprecationWarning"
markers = [  
    "integration: mark a test as an integration test",   
    "unit: mark a test as a unit test",       
    "e2e: mark a test as an end-to-end test",      
    "validation: mark a test as a validation test",               
] 
addopts = "--import-mode=importlib"

[tool.black]
line-length = 100
target_version = ["py311"]

[tool.isort]
profile = "black"
py_version = 311

[tool.flake8]
ignore = ['E501']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
